Namespace and project renamed
