+++ Task_DEV_6/Task_DEV_6/CustomStack.csproj
